+++ lib/QECCore/test/runtests.jl
@@ -1,0 +1,7 @@
+using QECCore
+using Test
+
+@testset "QECCore.jl" begin
+    @info "Testing QECCore.jl"
+    @test true
+end